Implement redis for relevant use cases

+++ server-side/controllers/bus.js
@@ -1,17 +1,57 @@
-const {Models} = require('../models');
+const { Models } = require('../models');
+const {getOrSetToRedis, clearRedis} = require('../utils/redis.js');
 
-const getBus = async(request,reply)=>{
+const getBus = async (request, reply) => {
+    const { busId } = request.params;
+    try {
+        const bus = await getOrSetToRedis(`bus:${busId}`, Models.Bus, async (Bus) => {
+            try {
+                const bus = await Bus.findOne({
+                    where: {
+                        busId: busId
+                    }
+                });
+                return bus;
+            } catch (error) {
+                console.log(error);
+                return;   
+            }
+        });
+
+        if (!bus) {
+            return reply('Error in getting bus').code(401);
+        }
+        return reply(bus).code(200);
+    } catch (error) {
+        console.log(error);
+        return reply('Internal server error').code(500);
+    }
+}
+
+const getBusDetails = async(request,reply)=>{
     const {busId} = request.params;
 
     try {
-        const bus = await Models.Bus.findOne({
-            where:{
-                busId: busId
+        const bus = await getOrSetToRedis(`bus:${busId}`, Models.Bus, async (Bus) => {
+            try {
+                const bus = await Bus.findOne({
+                    where: {
+                        busId: busId
+                    },
+                    include:[{
+                        model: Models.Operator,
+                        required: true
+                    }]
+                })
+                return bus;
+            } catch (error) {
+                console.log(error);
+                return;   
             }
         });
 
         if(!bus){
-            return reply('Error in getting bus').code(401);
+            return reply('No such bus!').code(401);
         }
 
         return reply(bus).code(200);
@@ -20,41 +60,14 @@ const getBus = async(request,reply)=>{
         return reply('Internal server error').code(500);
     }
 }
-const getOperatorBuses = async(request,reply)=>{
-    const {operatorId} = request.params
-    try {
-        const operator = await Models.Operator.findOne({
-            where:{
-                operatorId: operatorId
-            }
-        })
-        if(!operator){
-            return reply('Operator does not exist').code(401);
-        }
-        const buses = Models.Bus.findAll({
-            where:{
-                operatorId: operatorId
-            }
-        });
 
-        if(!buses){
-            return reply('No buses for this operator').code(401);
-        }
-
-        return reply(buses);
-    } catch (error) {
-        console.log(error);
-        return reply('Internal server error').code(500);
-    }
-}
-
-const createBus = async (request,reply)=>{
+const createBus = async (request, reply) => {
     const busDetails = request.payload;
 
     try {
         const newBus = await Models.Bus.create(busDetails);
 
-        if(!newBus){
+        if (!newBus) {
             return reply('Error in creating bus!').code(401);
         }
 
@@ -65,19 +78,28 @@ const createBus = async (request,reply)=>{
     }
 };
 
-const updateBus = async(request,reply)=>{
+const updateBus = async (request, reply) => {
     const updatedDetails = request.payload;
-    const {busId} = request.params;
+    const { busId } = request.params;
     try {
-        const bus = await Models.Bus.findOne({
-            where:{
-                busId: busId
+        const bus = await getOrSetToRedis(`bus:${busId}`, Models.Bus, async (Bus) => {
+            try {
+                const bus = await Bus.findOne({
+                    where: {
+                        busId: busId
+                    }
+                });
+                return bus;
+            } catch (error) {
+                console.log(error);
+                return;   
             }
         });
 
         if (!bus) {
             return reply('bus does not exist').code(401);
         }
+        clearRedis(`bus:${busId}`,Models.Bus);
         const updatedBuses = await Models.Bus.update(
             updatedDetails,
             {
@@ -86,11 +108,10 @@ const updateBus = async(request,reply)=>{
                 }
             }
         );
-        
-        if(!updatedBuses){
+
+        if (!updatedBuses) {
             return reply('Error updating bus details').code(401);
         }
-        
         return reply('Bus details updated successfully').code(200);
     } catch (error) {
         console.log(error);
@@ -98,17 +119,37 @@ const updateBus = async(request,reply)=>{
     }
 };
 
-const deleteBus = async(request,reply)=>{
-    const {busId} = request.params;
+const deleteBus = async (request, reply) => {
+    const { busId } = request.params;
     try {
-        const deletedBus = await Models.Bus.destroy({
-                where: {
-                    busId: busId
-                }
+        const bus = await getOrSetToRedis(`bus:${busId}`, Models.Bus, async (Bus) => {
+            try {
+                const bus = await Bus.findOne({
+                    where: {
+                        busId: busId
+                    }
+                });
+                return bus;
+            } catch (error) {
+                console.log(error);
+                return;   
             }
+        });
+
+        if(!bus){
+            return reply('No such bus').code(401);
+        }
+
+        clearRedis(`bus:${busId}`,Models.Bus);
+
+        const deletedBus = await Models.Bus.destroy({
+            where: {
+                busId: busId
+            }
+        }
         );
-        
-        if(!deletedBus){
+
+        if (!deletedBus) {
             return reply('Error deleting bus').code(401);
         }
         return reply('Bus details deleted').code(200);
@@ -120,7 +161,7 @@ const deleteBus = async(request,reply)=>{
 
 module.exports = {
     getBus,
-    getOperatorBuses,
+    getBusDetails,
     createBus,
     updateBus,
     deleteBus
